Add network id everywhere and get deploy test working
