+++ src/wallet/redux/reducers/__tests__/funding.test.ts
@@ -30,6 +30,7 @@ const defaults = {
   participants,
   privateKey: asPrivateKey,
   uid: 'uid',
+  networkId: 233,
 };
 
 const defaultsA = {
